requirements.txt
discord.py[voice] >= 2.2.0
yt-dlp
ipython
discord-racket >= 0.0.13